Fix: Fix `$env:TEMP` to use the full path instead of the MSDOS 8.3 shortened path for Windows users with long usernames
Previously, the value of `$env:TEMP` is, for a Windows user with a long username, something like `C:\users\somelo~1` for an actual path like `C:\users\somelongusername`. This value causes problems with `pandoc` which does not recognize shortened paths.

Now, the full path of `$env:TEMP` is resolved. This should fix the script for users with long usernames (anything more than 8 characters).

Useful links:
- https://superuser.com/questions/529400/how-does-progra1-path-notation-work
- https://web.archive.org/web/20131206010029/http://support.microsoft.com/kb/142982
- https://superuser.com/questions/348079/how-can-i-find-the-short-path-of-a-windows-directory-file
- https://superuser.com/questions/1524767/powershell-uses-the-short-8-3-form-for-envtemp

Fixes #167